rbstoreextendnotedetailviewcontroller: make the rotation test unsigned
Seven classes implement -shouldAutorotateToInterfaceOrientation: as the same
four instructions, `sub x8,x2,#1; cmp x8,#2; cset w0,cc`, which accepts only
the two portrait orientations. Six spell that as a pair of equality tests and
match. This one keeps the range-check idiom, which is closer to what the
binary does, but casts to NSInteger, and the comparison the binary makes is
unsigned: cc is carry-clear, an unsigned lower-than.

The difference shows at the unknown orientation, which is zero. Signed, zero
minus one is minus one and compares less than two, so the reconstruction
answers YES; the binary compares 0xFFFFFFFFFFFFFFFF against two and answers
NO. The cast is now NSUInteger on both sides of the comparison, which
reproduces the instruction exactly.

All seven are recorded as verified, along with the two ScoreData methods read
earlier. The checklist stands at 3143 of 6343.

+++ Project/RBStoreExtendNoteDetailViewController.m
@@ -390,7 +390,11 @@ static const UIViewAutoresizing kMaskFlexibleWidthTopBottom = 0x22; // W|Top|Bot
 }
 
 - (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)orientation {
-    return (NSInteger)(orientation - kFirstSupportedOrientation) < kSupportedOrientationCount;
+    // The range test is unsigned: the binary sets the result with `cset w0,cc` at 0x1a9308, and cc
+    // is carry-clear, an unsigned lower-than. Signed, the unknown orientation subtracts to -1 and
+    // compares less than two, so it would report YES where the binary reports NO.
+    return (NSUInteger)(orientation - kFirstSupportedOrientation) <
+           (NSUInteger)kSupportedOrientationCount;
 }
 
 #pragma mark - Content
